Fix eager batch collection not initialized
