+++ hibernate-core/src/main/java/org/hibernate/engine/spi/BatchFetchQueue.java
@@ -326,7 +326,7 @@ public class BatchFetchQueue {
 				}
 
 				if ( checkForEnd && i == end ) {
-					return resize( keys, i ); //the first key found after the given key
+					return keys; //the first key found after the given key
 				}
 
 				final boolean isEqual = collectionPersister.getKeyType().isEqual(
@@ -352,15 +352,7 @@ public class BatchFetchQueue {
 				}
 			}
 		}
-		return resize( keys, i ); //we ran out of keys to try
-	}
-
-	private Object[] resize(Object[] keys, int size) {
-		Object[] result = new Object[size];
-		for ( int i = 0; i < size; i++ ) {
-			result[i] = keys[i];
-		}
-		return result;
+		return keys; //we ran out of keys to try
 	}
 
 	private boolean isCached(Object collectionKey, CollectionPersister persister) {
