+++ hibernate-core/src/main/java/org/hibernate/metamodel/model/domain/internal/SetAttributeImpl.java
@@ -11,7 +11,6 @@ import java.util.Set;
 import org.hibernate.metamodel.internal.MetadataContext;
 import org.hibernate.metamodel.model.domain.SetPersistentAttribute;
 import org.hibernate.query.hql.spi.SqmCreationState;
-import org.hibernate.query.sqm.NodeBuilder;
 import org.hibernate.query.sqm.tree.SqmJoinType;
 import org.hibernate.query.sqm.tree.domain.SqmSetJoin;
 import org.hibernate.query.sqm.tree.from.SqmAttributeJoin;
