+++ hibernate-core/src/test/java/org/hibernate/orm/test/jpa/orphan/onetomany/DeleteSharedOneToManyOrphansTest.java
@@ -78,6 +78,7 @@ public class DeleteSharedOneToManyOrphansTest {
 
 	@Test
 	@TestForIssue(jiraKey = "HHH-11144")
+	@FailureExpected( jiraKey = "HHH-11144" )
 	public void testInitializingSecondCollection(EntityManagerFactoryScope scope) {
 		scope.inTransaction(
 				entityManager -> {
