+++ hibernate-core/src/test/java/org/hibernate/orm/test/manytomany/batchload/BatchedManyToManyTest.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.test.manytomany.batchload;
+package org.hibernate.orm.test.manytomany.batchload;
 
 import java.util.List;
 import java.util.Locale;
@@ -38,7 +38,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  * @author Steve Ebersole
  */
 @DomainModel(
-		xmlMappings = "org/hibernate/test/manytomany/batchload/UserGroupBatchLoad.hbm.xml"
+		xmlMappings = "org/hibernate/orm/test/manytomany/batchload/UserGroupBatchLoad.hbm.xml"
 )
 @SessionFactory(generateStatistics = true)
 @ServiceRegistry(
