+++ hibernate-core/src/test/java/org/hibernate/orm/test/manytomany/batchload/TestingBatchBuilder.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.test.manytomany.batchload;
+package org.hibernate.orm.test.manytomany.batchload;
 
 import org.hibernate.engine.jdbc.batch.internal.BatchBuilderImpl;
 import org.hibernate.engine.jdbc.batch.internal.NonBatchingBatch;
